% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_afvoershape.R
\name{create_afvoershape}
\alias{create_afvoershape}
\title{Subset afvoershape of waterbalance of interest}
\usage{
create_afvoershape(sf, waterbalance, check)
}
\arguments{
\item{sf}{shapefile containing all afvoergebieden (retrieve from data(afvoergebieden))}

\item{waterbalance}{code of waterbalance (e.g. KRWO_04_Kockengen)}
}
\description{
Subset afvoershape of waterbalance of interest
}
