% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seepage.R
\name{create_seepage}
\alias{create_seepage}
\title{Mask seepage raster to area waterbalance of interest (run create_afvoershape first)}
\usage{
create_seepage(rast, waterbalance)
}
\arguments{
\item{rast}{raster of LHM4.1 seepage}

\item{waterbalance}{code of waterbalance (e.g. KRWO_04_Kockengen)}
}
\description{
Mask seepage raster to area waterbalance of interest (run create_afvoershape first)
}
