% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_groundwaterlevels_timeseries.R
\name{extract_groundwaterlevels_timeseries}
\alias{extract_groundwaterlevels_timeseries}
\title{Extract groundwater levels as timeseries}
\usage{
extract_groundwaterlevels_timeseries(filepath, metingen_rownr, savedir)
}
\arguments{
\item{filepath}{filepath of waterbalance excel document}

\item{metingen_rownr}{amount of rows in sheet "metingen"}

\item{savedir}{directory to save results to}
}
\description{
Extract groundwater levels as timeseries
}
