% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_inoutlet_timeseries.R
\name{extract_inoutlet_timeseries}
\alias{extract_inoutlet_timeseries}
\title{Extract waterfluxes of in -and outlet}
\usage{
extract_inoutlet_timeseries(filepath, metingen_rownr, waterbalance)
}
\arguments{
\item{filepath}{filepath of waterbalance excel document}

\item{metingen_rownr}{amount of rows in sheet "metingen"}

\item{waterbalance}{code of waterbalance (e.g. KRWO_04_Kockengen)}
}
\description{
Extract waterfluxes of in -and outlet
}
