% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_other.R
\name{extract_other}
\alias{extract_other}
\title{Extract a range of other settings}
\usage{
extract_other(filepath, waterbalance)
}
\arguments{
\item{filepath}{filepath of waterbalance excel document}

\item{waterbalance}{code of waterbalance (e.g. KRWO_04_Kockengen)}
}
\description{
Extract a range of other settings
}
