% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stoffen_chloride.R
\name{extract_stoffen_chloride}
\alias{extract_stoffen_chloride}
\title{Extract concentrations of Cl as constants}
\usage{
extract_stoffen_chloride(filepath, metingen_rownr, waterbalance)
}
\arguments{
\item{filepath}{filepath of waterbalance excel document}

\item{metingen_rownr}{amount of rows in sheet "metingen"}

\item{waterbalance}{code of waterbalance (e.g. KRWO_04_Kockengen)}
}
\description{
Extract concentrations of Cl as constants
}
