% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_surfaceareas.R
\name{extract_surfaceareas}
\alias{extract_surfaceareas}
\title{Extract surface areas}
\usage{
extract_surfaceareas(filepath, savedir)
}
\arguments{
\item{filepath}{filepath of waterbalance excel document}

\item{savedir}{directory to save results to}
}
\description{
Extract surface areas
}
