% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_waterlevel_timeseries.R
\name{extract_waterlevel_timeseries}
\alias{extract_waterlevel_timeseries}
\title{Extract waterlevels as timeseries}
\usage{
extract_waterlevel_timeseries(filepath, metingen_rownr, waterbalance)
}
\arguments{
\item{filepath}{filepath of waterbalance excel document}

\item{metingen_rownr}{amount of rows in sheet "metingen"}

\item{waterbalance}{code of waterbalance (e.g. KRWO_04_Kockengen)}
}
\description{
Extract waterlevels as timeseries
}
