% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_waterlevelregime.R
\name{extract_waterlevelregime}
\alias{extract_waterlevelregime}
\title{Extract waterlevel regime}
\usage{
extract_waterlevelregime(filepath, waterbalance)
}
\arguments{
\item{filepath}{filepath of waterbalance excel document}

\item{waterbalance}{code of waterbalance (e.g. KRWO_04_Kockengen)}
}
\description{
Extract waterlevel regime
}
