% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_waterquality_timeseries.R
\name{extract_waterquality_timeseries}
\alias{extract_waterquality_timeseries}
\title{Extract measurements of Cl, N and P as timeseries}
\usage{
extract_waterquality_timeseries(filepath, metingen_rownr, waterbalance)
}
\arguments{
\item{filepath}{filepath of waterbalance excel document}

\item{metingen_rownr}{amount of rows in sheet "metingen"}

\item{waterbalance}{code of waterbalance (e.g. KRWO_04_Kockengen)}
}
\description{
Extract measurements of Cl, N and P as timeseries
}
