% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_param.R
\name{format_param}
\alias{format_param}
\title{Format extracted parameters to Artesia format}
\usage{
format_param(waterbalance)
}
\arguments{
\item{waterbalance}{acode of waterbalance (e.g. KRWO_04_Kockengen)}
}
\description{
Format extracted parameters to Artesia format
}
