% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_reeks.R
\name{format_reeks}
\alias{format_reeks}
\title{Format reeks to Artesia format}
\usage{
format_reeks(waterbalance)
}
\arguments{
\item{waterbalance}{acode of waterbalance (e.g. KRWO_04_Kockengen)}
}
\description{
Format reeks to Artesia format
}
