% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_series.R
\name{format_series}
\alias{format_series}
\title{Format series to Artesia format}
\usage{
format_series(waterbalance)
}
\arguments{
\item{waterbalance}{acode of waterbalance (e.g. KRWO_04_Kockengen)}
}
\description{
Format series to Artesia format
}
