% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_series_chloride.R
\name{format_series_chloride}
\alias{format_series_chloride}
\title{Format chloride timeseries to Artesia format}
\usage{
format_series_chloride(waterbalance)
}
\arguments{
\item{waterbalance}{code of waterbalance (e.g. KRWO_04_Kockengen)}
}
\description{
Format chloride timeseries to Artesia format
}
