% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_series_nitrogen.R
\name{format_series_nitrogen}
\alias{format_series_nitrogen}
\title{Format nitrogen timeseries to Artesia format}
\usage{
format_series_nitrogen(waterbalance)
}
\arguments{
\item{waterbalance}{code of waterbalance (e.g. KRWO_04_Kockengen)}
}
\description{
Format nitrogen timeseries to Artesia format
}
