% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_series_phosphorus.R
\name{format_series_phosphorus}
\alias{format_series_phosphorus}
\title{Format phosphorus timeseries to Artesia format}
\usage{
format_series_phosphorus(waterbalance)
}
\arguments{
\item{waterbalance}{code of waterbalance (e.g. KRWO_04_Kockengen)}
}
\description{
Format phosphorus timeseries to Artesia format
}
