% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_surfaceareas.R
\name{format_surfaceareas}
\alias{format_surfaceareas}
\title{Format extracted surface area data to Artesia format}
\usage{
format_surfaceareas(waterbalance)
}
\arguments{
\item{waterbalance}{code of waterbalance (e.g. KRWO_04_Kockengen)}
}
\description{
Format extracted surface area data to Artesia format
}
