% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_folderstructure.R
\name{set_folderstructure}
\alias{set_folderstructure}
\title{Set folder structure for extracting excel data}
\usage{
set_folderstructure(datafolder)
}
\arguments{
\item{datafolder}{folder where all excel waterbalances are located}
}
\description{
Set folder structure for extracting excel data
}
