% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_precipitation_timeseries.R
\name{extract_precipitation_timeseries}
\alias{extract_precipitation_timeseries}
\title{Extract precipitation and evaporation timeseries}
\usage{
extract_precipitation_timeseries(filepath, metingen_rownr, savedir)
}
\arguments{
\item{filepath}{filepath of waterbalance excel document}

\item{metingen_rownr}{amount of rows in sheet "metingen"}

\item{savedir}{directory to save results to}
}
\description{
Extract precipitation and evaporation timeseries
}
