% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pumpingstationcapacity.R
\name{extract_pumpingstationcapacity}
\alias{extract_pumpingstationcapacity}
\title{Extract pumping station capacity}
\usage{
extract_pumpingstationcapacity(filepath, waterbalance)
}
\arguments{
\item{filepath}{filepath of waterbalance excel document}

\item{waterbalance}{code of waterbalance (e.g. KRWO_04_Kockengen)}
}
\description{
Extract pumping station capacity
}
