% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_surfaceareas.R
\name{extract_surfaceareas}
\alias{extract_surfaceareas}
\title{Extract surface areas}
\usage{
extract_surfaceareas(filepath, waterbalance)
}
\arguments{
\item{filepath}{filepath of waterbalance excel document}

\item{waterbalance}{code of waterbalance (e.g. KRWO_04_Kockengen)}
}
\description{
Extract surface areas
}
